Attemt at adding animated widget
