+++ lib/timer_page.dart
@@ -1,5 +1,6 @@
 import 'package:flexify/app_state.dart';
 import 'package:flexify/main.dart';
+import 'package:flexify/timer_progress_widgets.dart';
 import 'package:flexify/utils.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
@@ -12,11 +13,6 @@ class TimerPage extends StatefulWidget {
 }
 
 class _TimerPageState extends State<TimerPage> {
-  String generateTitleText(Duration remaining) {
-    final minutes = (remaining.inMinutes).toString().padLeft(2, '0');
-    final seconds = (remaining.inSeconds % 60).toString().padLeft(2, '0');
-    return "$minutes:$seconds";
-  }
 
   @override
   Widget build(BuildContext context) {
@@ -29,14 +25,8 @@ class _TimerPageState extends State<TimerPage> {
       appBar: AppBar(
         title: const Text('Timer'),
       ),
-      body: Center(
-        child: Stack(
-          alignment: Alignment.center,
-          children: <Widget>[
-            progressWidget(context, duration, elapsed),
-            textWidget(context, timerState, remaining),
-          ],
-        ),
+      body: const Center(
+        child: TimerCircularProgressIndicator(),
       ),
       floatingActionButton: Visibility(
         visible: timerState.nativeTimer.isRunning(),
@@ -47,47 +37,5 @@ class _TimerPageState extends State<TimerPage> {
     );
   }
 
-  SizedBox progressWidget(BuildContext context, Duration duration, Duration elapsed) {
-    return SizedBox(
-      height: 300,
-      width: 300,
-      child: CircularProgressIndicator(
-        strokeCap: StrokeCap.round,
-        value: duration == Duration.zero ? 0 : elapsed.inMilliseconds / duration.inMilliseconds,
-        strokeWidth: 20,
-        backgroundColor:
-            Theme.of(context).colorScheme.onSurface.withOpacity(0.25),
-        valueColor: AlwaysStoppedAnimation<Color>(
-            Theme.of(context).colorScheme.primary),
-      ),
-    );
-  }
 
-  Column textWidget(BuildContext context, TimerState timerState, Duration remaining) {
-    return Column(
-      mainAxisAlignment: MainAxisAlignment.center,
-      children: <Widget>[
-        const SizedBox(height: 32.0),
-        Text(
-          generateTitleText(remaining),
-          style: TextStyle(
-            fontSize: 50.0,
-            color: Theme.of(context).textTheme.bodyLarge!.color,
-            fontWeight: FontWeight.bold,
-          ),
-        ),
-        TextButton(
-          onPressed: () {
-            final appState = Provider.of<AppState>(
-              context,
-              listen: false,
-            );
-            requestNotificationPermission();
-            timerState.addOneMinute();
-          },
-          child: const Text('+1 min'),
-        ),
-      ],
-    );
-  }
 }
